moved print to BrainGlobeAtlas
